1) This commit contains a lot of work necessary to work with the esp8266 module

+++ Firmware/header/time.h
@@ -67,7 +67,7 @@ void Time_startTimer(SoftTimerConfig timerConfig);
 uint32 Time_getTimerValue(SoftTimer timer);
 
 uint32 Time_getTimeOfday(void);
-uint32 Time_getSystemTime(void);
+uint64 Time_getSystemTime(void);
 boolean Time_isSecondBoundary(void);
 
 #endif //TIME_H
